Move the demo to its own file.
