+++ simtfl/demo.py
@@ -1,0 +1,69 @@
+"""
+A simple demo of message passing.
+"""
+
+from simpy import Environment
+
+from .message import PayloadMessage
+from .network import Network
+from .node import PassiveNode
+
+
+class Ping(PayloadMessage):
+    """
+    A ping message.
+    """
+    pass
+
+
+class Pong(PayloadMessage):
+    """
+    A pong message.
+    """
+    pass
+
+
+class PingNode(PassiveNode):
+    """
+    A node that sends pings.
+    """
+    def run(self):
+        """
+        (process) Sends a Ping message to every node.
+        """
+        for i in range(self.network.num_nodes()):
+            yield from self.send(i, Ping(i))
+            yield self.env.timeout(3)
+
+
+class PongNode(PassiveNode):
+    """
+    A node that responds to pings.
+    """
+    def handle(self, sender, message):
+        """
+        (process) Handles a Ping message by sending back a Pong message with the
+        same payload.
+        """
+        if isinstance(message, Ping):
+            yield self.env.timeout(5)
+            yield from self.send(sender, Pong(message.payload))
+        else:
+            yield from super().handle(sender, message)
+
+
+def run():
+    """
+    Runs the demo.
+    """
+    env = Environment()
+    network = Network(env, delay=4)
+    for i in range(10):
+        network.add_node(PongNode(i, env, network))
+
+    network.add_node(PingNode(10, env, network))
+
+    for i in range(network.num_nodes()):
+        env.process(network.start_node(i))
+
+    env.run()
