+++ simtfl/message.py
@@ -1,18 +1,12 @@
-class Message:
+class PayloadMessage:
     """
     A message with an arbitrary payload.
     """
     def __init__(self, payload):
         """
-        Constructs a `Message` with the given payload.
+        Constructs a `PayloadMessage` with the given payload.
         """
         self.payload = payload
 
     def __str__(self):
         return f"{self.__class__.__name__}({self.payload})"
-
-class Ping(Message):
-    pass
-
-class Pong(Message):
-    pass
